requirements.txt
fastapi
uvicorn[standard]
httpx
jsonschema-path
PyYAML
fastmcp
